Check if function call is of the right form before inspecting it.
The orderly_read feature scans the input file looking for calls to
`orderly.XXX` functions. Whenever it sees a function call it was
assuming it was in the form of an attribute lookup and would check if
the name of the part before the dot was equal to `orderly`.

This would crash if the method call was not on an attribute (eg. a
simple `print(...)` call), or if the part before the dot is not a simple
identifier (eg. a chained attribute lookup, or anything more
complicated).

This checks the type of the AST nodes more precisely to catch these
cases.