Change name of report scripts from orderly.py to <name>.py.
The will make outpack-py consistent with orderly2, which has recently
adopted this behaviour: [mrc-4969][mrc-4969] / [orderly2#122][orderly2#122].

Additionally, if we want to add the packet's directory to the search
path ([mrc-5227][mrc-5227]) having files called `orderly.py` would
shadow the actual orderly package. Using the name of the report is an
easy way to work around this.

[mrc-4969]: https://mrc-ide.myjetbrains.com/youtrack/issue/mrc-4969/Disambiguate-filenames
[mrc-5227]: https://mrc-ide.myjetbrains.com/youtrack/issue/mrc-5227/Reports-in-outpack-py-can-import-files-from-the-same-directory
[orderly2#122]: https://github.com/mrc-ide/orderly2/pull/122/files