Sort the list of IDs before comparing it.
The list of packet IDs returned by `unpacked()` is always sorted
lexicographically. Some tests were assuming that this was the same as
packet creation order, and were comparing the list of packets against
another value for equality.

However, if the packets were created very closely is time, the timestamp
part of their packet ID may be equal, in which case the relative order
is determined by the random bytes added to the packets. In that case,
lexicographical order may not match packet creation order. This is
especially likely on Windows where the timer resolution is quite coarse.