Allow reports to use multiprocessing module. (#42)
The multiprocessing module works in strange ways. It finds the top level
entrypoint of the program by looking for the `__main__` module, and then
re-executes it in each subprocess it creates. In each subprocess, the
entrypoint module is now called `__mp_main__`. To avoid infinite
recursive execution of the top program, the convention is to use an
`if __name__ == "__main__":` block surrounding the main code, such that it
will only run in the parent process, and not in the children.

We need to make sure that scripts written using this convention work
under orderly. The way to do this is to set `run_name="__main__"` when
invoking the script, and runpy sets everything up accordingly. Thanks to
this, multiprocessing will know to re-executes just the report script,
and not the top-level orderly code.

Additionally, when looking for calls to `orderly.parameters`, look
inside the `__name__` check, since that is where these need to be
located.