Return parameter values from the `orderly.parameters` call.
Previously we would inject parameter values as global variables before
the script starts executing. While this follows the behaviour of
orderly2, it confuses a lot of Python tooling, including linters and
type checkers. It also did not support executing orderly scripts outside
of a packet context (eg. interactively).

The `orderly.parameters` function now returns the parameter values as a
dictionary, which we can get from the global context. When running
outside of a packet context, it returns the parameters' default values.