Exclude pycache when copying packet inputs and outputs. (#45)
Python tends to produce a lot of `__pycache__` folders. These can be
found both in a report's source directory, if the user ran the
entrypoint script outside of an orderly context, and in the packet's
working directory after executing the packet.

There is never a good reason to include these in packets and we can
exclude them unconditionally, both when copying files in and out of a
packet's working directory.

As a side-effect of now using the same function to copy files in and out
of a packet, this commit also fixes the use of directory in a report's
source directory. Previously we were attempting to call `shutil.copy2`
on every entry, which would fail on directories.