Pull dependencies implicitly when running a report. (#47)
If a user runs a report with a `allow_remote=True` in their search
options, specifying a dependency will look for packets in remote
locations, and pull files from those locations as necessary. If
`pull_metadata` is set as well, the metadata will be synchronized from
all (enabled) locations.

By default, only the requested files are copied, and not whole packets.
However, if the `require_complete_tree` option is enabled on the root
then the entire packet, and all of its transitive dependencies, are
pulled as well.