Run reports in a separate Python process. (#41)
The runpy module we've been using to run reports provides very little
isolation. In particular, "any side effects (such as cached imports of
other modules) will remain in place after the functions have returned".
This means that if a user splits their report into multiple files, with
one file importing another, the latter will be cached and never reloaded
by Python, even if it is modified. Similarly, if the module was already
loaded before, the report runs against old code that does not match the
files included in the packet.

Using a separate subprocess gives us a much stronger isolation,
including all global variables and not sharing any of previously loaded
modules. It means a user can run a report, modify some of the imported
code, run the report again and have this behave as expected.

The sandbox is implemented by pickling the function and its argument
into a file, starting a new Python process which unpickles the file and
calls the target. Similarly, the function's return value, or any
exception it may throw, it pickled to an output file, which is read by
the parent.

I tried to use the multiprocessing module to implement this, but it puts
inconvenient requirements on the way the top-level source file is
written. The module re-evaluates the file in the subprocess, meaning its
contents must check for `__name__ == "__main__"` to avoid doing the work
again. It also doesn't work too well with the REPL. Using our own
implementation fixes these issues.

The Packet class is refactored a little and its scope is reduced to just
creating the packet and its metadata, without inserting it into the
repository. The insertion is now handled by a separate function. This is
done as a way of distinguishing what happens in the child process and
what happens in the parent.