Normalize paths in packet metadata. (#46)
The packet metadata keeps a record of what files are used and/or
produced by the report. When such files are directories, we want to
normalize the path separator. This will ensure that a packet produced on
Windows can be used on POSIX platforms and vice-versa.

Additionally, R (and therefore orderly2) uses forward slashes on all
platforms, including Windows. Using forward slashes in outpack-py means
packets are compatible and consistent across the two tools.

It's not obvious whether the functions that return copied paths to the
caller (eg. `orderly.shared_resource`) should return native paths or
POSIX paths. For now I've kept these as native paths.