Make root path absolute when opening.
For all user-facing operations that access a root, the user may pass in
an explicit location to the root. This path might be relative, in which
case it is assumed to be relative to the current working directory.

When running a report, we start a subprocess running inside a new
subdirectory of the drafts folder. This makes any existing relative
paths that we might pass to the packet builder meaningless in its
context, as the current working is now different. Since all these paths
are derived from the root path, the it is enough to transform that path
at the time when the root is opened.