Fix a bug when pulling and filtering at the same time.
If the search options had both `pull_metadata=True` and a `location`
filter set, the local location was excluded from the search regardless
of whether it was part of the filter.

This was caused by a call to the `location_resolve_valid` function with
`include_local` set to false, which was made as part of pulling the
metadata. This call would modify the filter in-place and remove "local"
from the list. Later, when the filter was used for the actual search,
local was always excluded.

The `location_resolve_valid` function now leaves the argument untouched
and instead always returns a copy of it.

Tangentially, the search feature now calls the `location_resolve_valid`
function as well, which will catch any unknown locations used in the
search options.