Add a new SSH-based location. (#48)
This is handy when running reports on the Linux cluster and the user
wants to pull the results back to their local machine.

The new location supports the same syntaxes as Git uses, either the
short SCP-like `username@hostname:path` syntax or the more explicit
URL-based `ssh://username@hostname/path`. The latter also allows
specifying a port number.

There's a bit of duplicated logic between the core implementation and
the new ssh location, since they both deal with the same file layout.
It felt like combining the two would require significant amount of
indirections by introducing a VFS that could be implemented both over
SFTP and locally, and wasn't worth it since the duplication isn't
excessive either.

The location pulling code got refactored a little bit. It used to make
many repeated calls to `driver.list()`, but that is inefficient when
these calls are over the network. While we could probably cache these
calls in the SSH driver, it is easy to just not make them.

The location driver's `fetch_file` interface is changed a little bit to
take more metadata in its arguments, including the packet name, id and
file path within the packet. This makes it much easier to find the file
when the remote location does not have a content-addressed file store.