Fix coverage measurement in subprocess.
When the sandbox was introduced it broke coverage measurement, as the
subprocess was not configured to enabled coverage.

While it is possible to setup coverage.py to track subprocesses,
switching to the pytest-cov plugin does it for us automatically. It also
takes care of combining results, so we don't need to do that as an
explicit step anymore.

Finally the `coverage.paths` section is removed since it was used in a
way that does not match what it actually does. We were listing
directories in our source for which we want coverage, when its actual
purpose is to defined equivalence classes.