Rename project to pyorderly.
Sadly the orderly name is already taken on PyPI. The repository was
named outpack, but this causes confusion as outpack refers to the
underlying storage abstraction.

While it is technically possible for Python packages to install multiple
modules with different names, it is quite unusual. Because of this, the
existing `outpack` module is folded under the `pyorderly` namespace as
`pyorderly.outpack`. I think it is useful to keep the distinction
between orderly and outpack.