Restrict the ways an outpack root can be opened.
We used to allow a `locate` argument that would control whether the
given root path is allowed to be a subdirectory of the actual root or
whether it must be the top-directory.

This pollutes the public interface for not much benefit. The behaviour
is replaced by always locating when the path is none (ie. search the
from the current working directory upwards) and never location when an
explicit path is given.

This matches the change made to the R package in
https://github.com/mrc-ide/orderly/pull/183.