Use direct device flow authentication.
We replaced GitHub authentication in Packit with a native device flow
authentication. Instead of authenticating against GitHub and then
trading a GitHub PAT for a Packit token, pyorderly can now authenticate
directly against Packit.

This simplifies the implementation and allows us to login to Packit
instances that do not use GitHub authentication (such as Montagu).

This flow has been in use in the R orderly package for a while now and
is working quite well.